User module changed user controllers
