+++ src/controllers/mainController.js
@@ -1,10 +1,9 @@
-
 const mainController = {};
 
 mainController.get = (req, res) => {
-    return res.json({
-        data: 'This is a full stack app!'
-    });
+  return res.json({
+    data: "This is a full stack app!",
+  });
 };
 
 module.exports = mainController;
